fix(persistence): widen counters and market prices
Why:
- player periodic flush was already failing on int32 overflow
- market unitPrice allows values far above PostgreSQL integer range
- protobuf attr specialStats still capped foundation/combatExp at uint32

What:
- widen legacy player cumulative counters and user online seconds to bigint
- widen market order/trade unitPrice to bigint
- add startup column-capacity guards for legacy users/players/market tables
- upgrade AttrUpdate protobuf specialStats foundation/combatExp to uint64
- sync server-next legacy smoke schemas with the widened legacy columns

Checks:
- pnpm --filter @mud/shared build
- pnpm --filter @mud/server build
- node --check packages/server-next/src/tools/next-auth-bootstrap-smoke.js
- node --check packages/server-next/src/tools/compat/legacy-auth-smoke.js

+++ packages/shared/src/network-protobuf.ts
@@ -253,8 +253,8 @@ message AttrUpdatePayload {
 }
 
 message PlayerSpecialStatsPayload {
-  optional uint32 foundation = 1;
-  optional uint32 combatExp = 2;
+  optional uint64 foundation = 1;
+  optional uint64 combatExp = 2;
 }
 
 message AttributesPayload {
